+++ src/main/java/com/getfei/jobSpider/service/ITechnologyService.java
@@ -28,6 +28,6 @@ public interface ITechnologyService {
 	/**新增别名*/
 	MongoResult addAlias(String name,String type,String alias);
 	
-	long count();
+	Integer count();
 	
 }
